Don't evaluate attrs to be overridden
The Factory#attrs method first evaluates all params as per the initialiser object
it then overrides them with passed in extra attrs.

This causes unnecessary evaluation for things that we are going to override any way.
For example, if there is an association and an existing object is passed in to override the
association, Factory#attrs will still create the association only to replace it with the
passed object

This commit filters out attrs that are going to be overridden and evaluates only the
required ones.